Changed the strategy for changing pages. Earlier it was located in PageView, but since we needed to handle close(), transitions etc gracefully and the necessity to think about future requirements during Region.show(), i decided to go for implementation with Region.showPage(), openPage() and closePage() explicityly for PageViews.
Decided not to allow data-transition either in links or in page markups because that is not how developers work with BB. In BB, links are not auto bound, intead, they are handled by methods in the view or router, so it is not necessary to declaratively specify transition in the TL structure. Instead, an api Region.showPage(view, options) will take transition options when called.

Decided to rule out dialog for the time being. Check data-attribute.txt for more info.

Decided to rule out data-rel because auto link binding is not required in BB app